% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi.R
\name{bayes_ridge_semi}
\alias{bayes_ridge_semi}
\title{Bayesian ridge for longitudinal semiparemetric models}
\usage{
bayes_ridge_semi(
  y,
  grp,
  Bmat,
  Xmat,
  Kmat = NULL,
  dim_block = NULL,
  ranef = NULL,
  burn = 0,
  size = 1000,
  init = NULL,
  prior = NULL,
  prec = NULL,
  debug = TRUE
)
}
\arguments{
\item{y}{A vector of the response vector.}

\item{grp}{A list of two numeric/factor/character vectors, grp$pop and
grp$sub, specifying the population and subject to which each observation
belongs.}

\item{Bmat}{A list of two design matrices for the splines, Bmat$pop and
Bmat$sub for population- and subject-specific curves respectively.}

\item{Xmat}{A design matrix for the non-spline terms, i.e. fixed and random
effects.}

\item{Kmat}{The penalty matrix applied on the population splines
coefficients.}

\item{dim_block}{The number of subject spline coefficients to be treated as
correlated. The first `dim_block` of the coefficients are correlated. Set
this to 0 if all subject coefficients are uncorrelated, or `dim(Bmat$sub)`
if everything is correlated.}

\item{ranef}{A numeric vector of indices of the Xmat columns where the
corresponding coefficients should be treated as random effects. NULL if
everything is fixed. This option will get overridden by the 'beta' term in
`prec`.}

\item{burn}{The number of samples to be burned before actual sampling.}

\item{size}{The number of samples to be obtained from the samples.}

\item{init}{An initial values of the regression coefficients for the
sampler...}

\item{prior}{unknown yet}

\item{prec}{An optional list of precisions with 'theta', 'delta', 'beta' and
'eps'. If supplied, the precisions of the model are fixed. All the elements
are square matrices, except 'eps' which is a number. The 'beta' term here
will override `ranef` (i.e. all the 0 in diagonal will result in the
corresponding beta being fixed).}
}
\value{
A list of length two, posterior samples and means are stored in
  'samples' and 'means' elements respectively. Within each elements, 'coef'
  are the regression coefficients and 'prec' the precisions (i.e. the inverse
  of variances).

 The samples are organised as arrays, except coef$beta (matrix) and prec$eps
 (vector). The samples are always populated along the last dimension of the
 array, matrix or vector. For the 'coef' samples, the first dimension of the
 array/matrix is always the dimension of the coefficients, followed by the
 indices of populations/subjects (for 'theta' and 'delta' only). All the
 precisions except 'eps' are symmetrical square matrices.
}
\description{
This is a Gibbs sampler v3 for fitting Bayesian longitudinal semiparametric
models. It assumes that there are multiple populations in the model and each
population is modelled by a 'population' curve. Within each population, they
are multiple subjects, and each subject are modelled by a 'subject'
curve. The subject curves are treated as deviations from their respective
population curves. On top of that, fixed or random effects can be added to
the model. This is useful when, for example, a particular treatment was
applied to some of the populations or subjects, and the user is interested in
the effect of the treatment.
}
\details{
The population and subject curves are modelled as linear (in statistical
sence, not only stright lines). This includes polynomials and splines.

The mathematical model is

y_i = Bmat$pop_i %*% \theta_i + Bmat$sub_i %*% \delta_i + Xmat_i %*% \beta +
\epsilon_i

where 'i' is the index of subjects, y_i is a vector of observation for the
i^{th} subjectm, Bmat$pop_i and Bmat$sub_i are the model/design matrices for
the population and subject curves (of the i^{th} subject) respectively, and
\epsilon_i are Gaussian errors. The \theta_i should be intepreted as the
regression coefficients of the population curve of the i^{th} subject, and
the subjects belonging to the same population will have the same \theta. This
implies that some of the \theta_i will be identical. However, \delta_i will
be different for each subject, while \beta will be the same for all subjects.

The coefficients \theta, \delta, \beta and \epsilon are all Gaussian, but
their covariance structures are not necessarily diagonal. Consult the
manual/paper for more info.

The sampler first updates the precision, then the coefficients. The joint
conditional posterior of all the coefficients is often highly correlated, but
fortunately a closed-form expression is available and is utilised here. This
implies that the convergence of this Gibbs sampler is quick and does not
require burning many samples in the beginning.
}
