% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi2.R
\name{bayes_ridge_semi_v4}
\alias{bayes_ridge_semi_v4}
\title{Bayesian ridge for longitudinal semiparemetric models}
\usage{
bayes_ridge_semi_v4(
  y,
  Bmat,
  Xmat,
  prior = NULL,
  prec = NULL,
  init = NULL,
  burn = 0,
  size = 1000,
  debug = TRUE
)
}
\arguments{
\item{y}{A vector of the response vector.}

\item{Bmat}{A named list of design matrices for the splines terms with the
following attributes: 'spl_dim', 'is_sub', 'level'. For subject-specific
splines (i.e. is_sub = TRUE), it should also have 'block_dim' and
'index'. For other splines, it should have 'penalty'. At least an element
of Bmat must be 'is_sub = TRUE'. See details.}

\item{Xmat}{A list of design matrices for the non-spline terms, e.g. fixed
and random effects.}

\item{prior}{unknown yet}

\item{prec}{An optional list of precisions with 'theta', 'delta', 'beta' and
'eps'. If supplied, the precisions of the model are fixed. All the elements
are square matrices, except 'eps' which is a number. The 'beta' term here
will override `ranef` (i.e. all the 0 in diagonal will result in the
corresponding beta being fixed).}

\item{init}{An initial values of the regression coefficients for the
sampler...}

\item{burn}{The number of samples to be burned before actual sampling.}

\item{size}{The number of samples to be obtained from the samples.}
}
\value{
A list of length two, posterior samples and means are stored in
  'samples' and 'means' elements respectively. Within each elements, 'coef'
  are the regression coefficients and 'prec' the precisions (i.e. the inverse
  of variances).

 The coef samples are organised as a list of arrays for splines and a list of
 matrices for effects. (rewrite the precision part, except coef$beta (matrix)
 and prec$eps (vector)). The samples are always populated along the last
 dimension of the array, matrix or vector. For the 'coef' samples, the first
 dimension of the array/matrix is always the dimension of the coefficients,
 followed by the indices of populations/subjects (for 'theta' and 'delta'
 only). All the precisions except 'eps' are symmetrical square matrices.
}
\description{
This is a Gibbs sampler v3 for fitting Bayesian longitudinal semiparametric
models. It assumes that there are multiple populations in the model and each
population is modelled by a 'population' curve. Within each population, they
are multiple subjects, and each subject are modelled by a 'subject'
curve. The subject curves are treated as deviations from their respective
population curves. On top of that, fixed or random effects can be added to
the model. This is useful when, for example, a particular treatment was
applied to some of the populations or subjects, and the user is interested in
the effect of the treatment.
}
\details{
The population and subject curves are modelled as linear (in statistical
sence, not only stright lines). This includes polynomials and splines.

'index' is a list of numeric vector specifying the position of rows for each
  level of the factor. The precision of the prior of the regression
  coefficient is up to a multiplicative constant of crossprod('penalty').

 The mathematical model is

y_i = Bmat$pop_i %*% \theta_i + Bmat$sub_i %*% \delta_i + Xmat_i %*% \beta +
\epsilon_i

where 'i' is the index of subjects, y_i is a vector of observation for the
i^{th} subjectm, Bmat$pop_i and Bmat$sub_i are the model/design matrices for
the population and subject curves (of the i^{th} subject) respectively, and
\epsilon_i are Gaussian errors. The \theta_i should be intepreted as the
regression coefficients of the population curve of the i^{th} subject, and
the subjects belonging to the same population will have the same \theta. This
implies that some of the \theta_i will be identical. However, \delta_i will
be different for each subject, while \beta will be the same for all subjects.

The coefficients \theta, \delta, \beta and \epsilon are all Gaussian, but
their covariance structures are not necessarily diagonal. Consult the
manual/paper for more info.

The sampler first updates the precision, then the coefficients. The joint
conditional posterior of all the coefficients is often highly correlated, but
fortunately a closed-form expression is available and is utilised here. This
implies that the convergence of this Gibbs sampler is quick and does not
require burning many samples in the beginning.
}
