% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{combine_array}
\alias{combine_array}
\title{Combine vectors or arrays recursively}
\usage{
combine_array(...)
}
\arguments{
\item{...}{vectors, arrays or lists.}
}
\value{
A combined vector/array, or a list of combined vectors/arrays. The
  structured of the list is preserved.
}
\description{
Combine vectors or arrays given in .... If ... are (nested) lists, then it
recursively goes down the list to combine vectors/arrays. The structure of
the list is preserved.
}
