% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{combine_fm}
\alias{combine_fm}
\title{Combine multiple flexss objects into one.}
\usage{
combine_fm(...)
}
\arguments{
\item{...}{a list of flexss objects, or flexss objects themselves}
}
\value{
a flexss object with all the object properties (except samples and
  means) inherited from the first flexss object. The posterior means are
  recalculated from the combined samples.
}
\description{
Combine the samples of multiple flexss objects and chuck them into the
'samples' field of the first flexss object. This is useful when running
multiple MCMC chains simulatenously and combining them together.
}
