% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ridge.R
\name{fit_bs_splines_v3}
\alias{fit_bs_splines_v3}
\title{Fit a B-spline model with population and subject-specific curves.}
\usage{
fit_bs_splines_v3(
  data,
  K,
  deg,
  Xmat,
  ranef,
  size,
  burn,
  ridge = FALSE,
  init = NULL,
  prior = NULL
)
}
\arguments{
\item{data}{A data frame with at least three columns (\code{x}, \code{y} and
\code{sub}). If a \code{pop} column is also given, separate mean curves
are fitted to each population.}

\item{K}{A list of numbers of interior knots for the population and subject
curves. The list must contain \code{sub} and \code{pop}.}

\item{deg}{The degree of the spline polynomial}

\item{Xmat}{A design matrix for the non-spline terms, i.e. fixed and random
effects.}

\item{ranef}{A numeric vector of indices of the Xmat columns where the
corresponding coefficients should be treated as random effects. NULL if
everything is fixed. This option will get overridden by the 'beta' term in
`prec`.}

\item{size}{The number of samples to be drawn from the posterior.}

\item{burn}{The number of samples to burn before recording. Default to
one-tenth of \code{size}.}

\item{ridge}{Use ridge regression or linear mixed effect models?}

\item{init}{List of matrices of the coefficients, containing \code{sub} and
\code{pop}. The columnns of the matrices correspond to each
population/subject.}

\item{prior}{List of hyperparameters of the covariance priors.}
}
\value{
A list with posterior means, samples and information of the basis
    functions.
}
\description{
Fit a B-spline model which gives population and subject-specific curves.
}
\details{
This model is fitted in a Bayesian framework, that is, this routine gives the
samples drawing from the posterior distribution associated to each regression
coefficients. The population (mean) curves are penalised with a difference
matrix of order \code{deg + 1}. The knots are equally spaced between
\code{range(data$x)}.
}
