% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ridge.R
\name{fit_bs_splines_v4}
\alias{fit_bs_splines_v4}
\title{Fit a B-spline model with population and subject-specific curves.}
\usage{
fit_bs_splines_v4(
  fixed,
  data,
  spline,
  random = NULL,
  size = 1000,
  burn = 0,
  ridge = FALSE,
  init = NULL,
  prior = NULL,
  prec = NULL
)
}
\arguments{
\item{fixed}{A formula y ~ x1 + x2 ... specifying the response and fixed
effects. This routine looks up the variable from the data, before looking
at the environment where the formula is defined.}

\item{data}{A data frame with at least three columns (\code{x}, \code{y} and
\code{sub}). If a \code{pop} column is also given, separate mean curves are
fitted to each population.}

\item{spline}{A list of formulae of splines s(x, by =, K = , deg = ,
...). This routine looks up the variable from the data, before looking
at the environment where the formula is defined. See `s` for more details.}

\item{random}{A formula ~ u1 + u2 ... specifying the random effect term.}

\item{size}{The number of samples to be drawn from the posterior.}

\item{burn}{The number of samples to burn before recording. Default to
one-tenth of \code{size}.}

\item{ridge}{Use ridge regression or linear mixed effect models?}

\item{init}{List of matrices of the coefficients, containing \code{sub} and
\code{pop}. The columnns of the matrices correspond to each
population/subject.}

\item{prior}{List of hyperparameters of the covariance priors.}
}
\value{
A list with posterior means, samples and information of the basis
    functions.
}
\description{
Fit a B-spline model which gives population and subject-specific curves.
}
\details{
This model is fitted in a Bayesian framework, that is, this routine gives the
samples drawing from the posterior distribution associated to each regression
coefficients. The population (mean) curves are penalised with a difference
matrix of order \code{deg + 1}. The knots are equally spaced between
\code{range(data$x)}.

The intercepts for all splines, other than the subject-specific splines, are
removed by default. This is to ensure that the intercept term in the spline
is not spanning the same space as the (potential) main effect terms (and
consequently having an unidentifiable model). For subject-specific splines,
all the coefficients are penalised, so including an intercept will not result
in an unidentifiable model.

 Therefore, the intercept should be added manually either by setting
'intercept = TRUE' or adding it as fixed effects.
}
