% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{flatten_chain}
\alias{flatten_chain}
\title{Flatten a flexss object into a matrix.}
\usage{
flatten_chain(fm)
}
\arguments{
\item{...}{A flexss object.}
}
\value{
A matrix of samples.
}
\description{
Convert the samples in a flexss object into a n_samples * n_parameters
matrix.
}
