% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{flatten_chains}
\alias{flatten_chains}
\title{Flatten flexss objects into a 3D-array.}
\usage{
flatten_chains(...)
}
\arguments{
\item{...}{A list of flexss objects, or flexss objects themselves}
}
\value{
A 3D array of samples.
}
\description{
Convert the samples in multiple flexss objects into a n_samples * n_objects *
n_parameters 3D-array.
}
