% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ridge.R
\name{get_design_bs}
\alias{get_design_bs}
\title{Construct an equidistant B-splines design matrix}
\usage{
get_design_bs(x, K, deg, EPS = 1e-06)
}
\arguments{
\item{x}{predictor vector. Required.}

\item{K}{number of inner knots, or a vector of all knots (interior,
boundaries, and knots outside extrema). Required.}

\item{deg}{the degree of polynomial which the B-splines span. Required.}

\item{EPS}{tolerance error.}
}
\value{
a list with components `design' (\code{length(x)} by \code{K + deg +
  1} design matrix) and all `knots' (interior and boundaries knots, and knots
  outside extrema). If type == 'LMM', the transformation matrix is returned
  in the 'transform' field.
}
\description{
\code{get_design_bs} returns an equidistant B-splines design matrix without
explicitly specifying the location of the knots.
}
\details{
The knots locations are the minimum and maximum of \code{x}, and \code{K}
equidistant points between these two extrema. The B-splines are equidistant,
i.e. no multiple knots at both ends of \eqn{x}. This function uses
\code{splines::splineDesign}.
}
