% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-ridge.R
\name{get_diff_mat}
\alias{get_diff_mat}
\title{Construct a difference matrix}
\usage{
get_diff_mat(size, k)
}
\arguments{
\item{size}{the column size of the difference matrix. Required.}

\item{k}{the order of difference. Required.}
}
\value{
A \code{size - k} by \code{size} matrix.
}
\description{
\code{get_diff_mat} returns a difference matrix of an arbitrary order and size.
}
\details{
This function returns a \eqn{k^{th}} order difference matrix \eqn{D}, such
that \eqn{Dx} gives a vector of \eqn{k^{th}} order differences of \eqn{x}.
The parameter \code{size} is usually the dimension of \eqn{x}.
}
