% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi2.R
\name{get_para_v4}
\alias{get_para_v4}
\title{Get dims and names of the spline and effect terms.}
\usage{
get_para_v4(Bmat, Xmat)
}
\arguments{
\item{Bmat, Xmat}{Design matrices of the spline and effect terms.}
}
\value{
A list of dims and levels
}
\description{
Get dims and names of the spline and effect terms.
}
