% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi2.R
\name{is_match_list}
\alias{is_match_list}
\title{Check list structure}
\usage{
is_match_list(...)
}
\arguments{
\item{...}{Lists to be checked.}
}
\value{
Boolean
}
\description{
Check if all the lists have the same structure.
}
