% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi2.R
\name{prec_to_cov}
\alias{prec_to_cov}
\title{Convert precision to variance/covariance}
\usage{
prec_to_cov(prec)
}
\arguments{
\item{prec}{a vector of precision, an array of precision matrices}
}
\value{
variance or covariance matrix
}
\description{
Convert precision (matrices) to variance (covariance matrices).
}
