% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{rhat}
\alias{rhat}
\title{Rhat convergence diagnostic}
\usage{
rhat(sims)
}
\arguments{
\item{sims}{A 2D array _without_ warmup samples (# iter * # chains).}
}
\value{
A single numeric value for the effective sample size.
}
\description{
Compute Rhat convergence diagnostic as the maximum of rank normalized
split-Rhat and rank normalized folded-split-Rhat for one parameter.
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2019). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{arXiv preprint} \code{arXiv:1903.08008}.
}
