% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{u_scale}
\alias{u_scale}
\title{Rank uniformization}
\usage{
u_scale(x)
}
\arguments{
\item{x}{A numeric array of values.}
}
\value{
A numeric array of rank uniformized values with the same
    size as input.
}
\description{
Compute rank uniformization for a numeric array. First replace each
value by its rank. Average rank for ties are used to conserve the
number of unique values of discrete quantities. Second, uniformize
ranks to scale [1/(2S), 1-1/(2S)], where S is the the number of values.
}
