% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi2.R
\name{update_coefs_v4}
\alias{update_coefs_v4}
\title{Update coefficients}
\usage{
update_coefs_v4(y, Bmat, Xmat, xBs, kprec, debug = FALSE)
}
\arguments{
\item{y}{An n by 1 column matrix of the responce}

\item{Bmat}{A list of spline term design matrices. The term for subject
curves must be at the front. Required attributes of each matrix: spl_dim
(num), level (char), is_sub (bool), penalty (matrix, non-sub spline only),
block_dim (num, sub spline only)}

\item{Xmat}{}
}
\value{
precision matrix for the coefs
}
\description{
Update coefficients
}
