% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi2.R
\name{bayes_ridge_semi_v4}
\alias{bayes_ridge_semi_v4}
\title{Bayesian ridge for longitudinal semiparemetric models}
\usage{
bayes_ridge_semi_v4(
  y,
  Bmat,
  Xmat,
  prior = NULL,
  prec = NULL,
  init = NULL,
  burn = 0,
  size = 1000,
  debug = TRUE
)
}
\arguments{
\item{y}{A vector of the response vector.}

\item{Bmat}{A named list of design matrices for the splines terms (or a list
of design matrices of each subject for the subject spline) with at least
the following attributes: 'spl_dim', 'is_sub', 'level'. For
subject-specific splines (i.e. is_sub = TRUE), it is a list of matrices and
should also have 'block_dim' and 'index'. For other splines, it is a matrix
and should have 'penalty'. At least an element of Bmat must be 'is_sub =
TRUE'. See details.}

\item{Xmat}{A named list of design matrices for the non-spline terms,
e.g. fixed and random effects.}

\item{prior}{unknown yet}

\item{prec}{An optional named list with three elements 'spl', 'eff' and
'eps'. If supplied, the precisions of the model are fixed. Each of the
'spl' and 'eff' terms is a named list of precisions that correspond to
'Bmat' and 'Xmat' respectively, and the names of the list must match. All
the elements are square matrices with the dimension matching 'spl_dim' of
Xmat or Bmat, except 'eps' which is a number. A zero precision will imply
that the term is fixed.}

\item{init}{An initial values of the regression coefficients for the
sampler...}

\item{burn}{The number of samples to be burned before actual sampling.}

\item{size}{The number of samples to be obtained from the samples.}
}
\value{
A list of length two, posterior samples and means are stored in
  'samples' and 'means' elements respectively. Within each elements, 'coef'
  are the regression coefficients and 'prec' the precisions (i.e. the inverse
  of variances).

 The coef samples are organised as a list of arrays for splines and a list of
 matrices for effects. (rewrite the precision part, except coef$beta (matrix)
 and prec$eps (vector)). The samples are always populated along the last
 dimension of the array, matrix or vector. For the 'coef' samples, the first
 dimension of the array/matrix is always the dimension of the coefficients,
 followed by the indices of populations/subjects (for 'theta' and 'delta'
 only). All the precisions except 'eps' are symmetrical square matrices.
}
\description{
This is a Gibbs sampler v3 for fitting Bayesian longitudinal semiparametric
models. It assumes that there are multiple populations in the model and each
population is modelled by a population 'mean' curve. Within each population,
they are multiple subjects, and each subject are modelled by a 'subject'
curve. The subject curves are treated as deviations from their respective
mean curves. On top of that, fixed or random effects can be added to the
model. This is useful when, for example, a particular treatment was applied
to some of the populations or subjects, and the user is interested in the
effect of the treatment.
}
\details{
The mean and subject curves are modelled as linear (in statistical sence, not
only stright lines). This includes polynomials and splines. The attributes of
their design matrices (i.e. Bmat) are

'spl_dim' is the dimension of the splines.

'is_sub' is a boolean that specifies whether the spline term is a
subject-specific deviations.

'level' is the name of each population or subject. NA if there is only one
population.

'index' is a list of numeric vector specifying the position of rows for each
  subject. Only applicable for subject curves.

'block_dim' is the dimension of the subject deviations where the precision
should be considered as a block. See paper for more details.

'penalty' relates to the precision of the prior of the mean curve
  coefficient, which is a multiple of crossprod('penalty'), i.e. ('penalty'
  %*% \theta_i) is shrunk to 0.

The semiparametric model is

y = Bmat[[1]] %*% \theta_1 + ... + Xmat[[1]] %*% \beta_1 + ... + Bmat[[sub]]
%*% \delta + \epsilon

y is a vector of observation, Bmat[[...]] %*% \theta are the splines that
constitute the mean curve, Xmat[[...]] %*% \beta are the fixed/random
effects, Bmat[[sub]] %*% \deltais the subject-specific deviations, and
\epsilon are Gaussian errors.

The coefficients \theta, \delta, \beta and \epsilon are all Gaussian, but
their covariance structures are not necessarily diagonal. Consult the
manual/paper for more info.

The sampler first updates the coefficients, then the precisions. The joint
conditional posterior of all the coefficients is often highly correlated, but
fortunately a closed-form expression is available and is utilised here. This
implies that the convergence of this Gibbs sampler is quick and does not
require burning many samples in the beginning. The starting values of the
sampler are precisions with a reasonably large value (hence a large penalty).
}
