% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ridge.R
\name{predict.fsso}
\alias{predict.fsso}
\title{Predict with posterior mean}
\usage{
\method{predict}{fsso}(object, newdata = NULL, level = NULL)
}
\arguments{
\item{object}{an 'fsso' object.}

\item{newdata}{a data frame with all the covariate required for prediction.}

\item{level}{a string specify the name of the spline. Only the component
correponding to that spline is returned together with the effect terms).}
}
\description{
Return prediction with the posterior mean of the regression coefficients.
}
