% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-ridge.R
\name{s}
\alias{s}
\title{Return the design matrix of a spline}
\usage{
s(
  x,
  by = NULL,
  knots = NULL,
  degree = NULL,
  intercept = NULL,
  type = "bs",
  is_sub = FALSE,
  block_dim = NULL
)
}
\arguments{
\item{x}{a variable of the design points.}

\item{by}{a grouping variable, or if NULL, a common spline is fitted.}

\item{knots}{the number of internal knots.}

\item{degree}{the degree of spline.}

\item{intercept}{TRUE/FALSE whether an intercept should be included in the
spline.}

\item{type}{the type of basis function. Currently only B-spline ('bs') is
supported.}

\item{is_sub}{is this the subject-specific curves?}

\item{block_dim}{the (first) number of terms in the subject-specific curves
that should be treated as a block, rather than i.i.d.}
}
\description{
Return the design matrix of a spline.
}
\details{
For non-subject-specific splines, the first 'degree + 1' terms correspond to
the coefficients of a polynomial of 'degree' and are not penalised. The rest
of the coefficients are shrunk to zero so that when all of these terms are
zero, the fit is a polynomial.

Since the polynomial (which includes lines parallel to the x-axis) terms are
not penalised, the intercept term in the polynomial should be removed to
avoid spanning the same space as the (potential) main effect terms. For
subject-specific splines, all the coefficients are penalised, so including an
intercept will not result in an unidentifiable model.

Due to the way which B-spline basis is parameterised and transformed, it is
advisable to treat the first 'degree + 1' coefficients of the
subject-specific spline as correlated Gaussian random effects (i.e. with full
covariance matrix). The rest of the spline coefficients (subject-specific or
otherwise) are independent Gaussian. 'block_dim' is only applicable for the
subject-specific spline.
}
